chore: remove duplicate code in simulations
